Perf: static playout-cap counts (4x effective batch) + bf16 sample storage
- playout_cap: exactly round(full_prob*T) full-search steps per rollout at
  random positions; ONLY those are stored. The per-step bernoulli produced
  ~75% weight-0 buffer rows, silently diluting every training batch 4x.
  The weight field is gone (all stored rows train).
- obs + policy targets stored as bf16 (obs values and softmax targets are
  bf16-safe; CE accumulates in f32): halves the two largest scan/replay
  tensors -> restores 1024 games/chip within v5e HBM.

+++ jaxarimaa/trainer.py
@@ -77,7 +77,8 @@ def loss_fn(params, apply_fn, batch, value_weight, aux_weights=(0.0, 0.0, 0.0)):
     obs = batch["obs"]
     logits, value, aux = jax.vmap(lambda o: apply_fn(params, o))(obs)
     logp = jax.nn.log_softmax(logits, axis=-1)
-    policy_loss = -jnp.sum(batch["policy_target"] * logp, axis=-1)
+    # targets may be stored bf16 in the replay buffer; accumulate the CE in f32
+    policy_loss = -jnp.sum(batch["policy_target"].astype(jnp.float32) * logp, axis=-1)
     value_loss = (value - batch["value_target"]) ** 2
     # Per-sample weights (playout-cap: fast moves have weight 0). Default weight 1 =>
     # weighted mean is the plain mean, so baseline behavior is unchanged.
